backend/requirements.txt
aiohttp
bcrypt
fastapi[standard]
langchain-community
langchain-openai
passlib
psycopg2-binary
pydantic[email]
python-dotenv
python-jose
python-multipart
requests
rich
sqlalchemy
websockets
livekit-agents
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-elevenlabs
livekit-plugins-silero
livekit.plugins.turn_detector
pytz
google-cloud-storage